// angular_client/src/app/shared/enums/routerPaths.enum.ts
export enum RouterPathsEnum {
      DEFAULT = "",
      LOGIN = "login",
      REGISTER = "register",
      OVERVIEW = "overview",
      ANALYTICS = "analytics",
      MAILING = "emails",
      HISTORY = "history",
      ORDER = "order",
      CATEGORIES = "categories",
      PROFILE = "profile",
      NEW = "new",
      ID = ":id",
}